# api/src/generated/openapi.yaml
openapi: 3.0.0
components:
...
  schemas:
...
    GroupScoped:
      description: Group scoped by membership - API returns only appropriate fields
        per user's status
      properties:
        id:
          type: string
        name:
          type: string
        desc:
          type: string
        thumbnail:
          type: string
          nullable: true
        avatarSeed:
          type: string
          nullable: true
        isPublic:
          type: boolean
        memberCount:
          type: number
          format: double
        membershipStatus:
          $ref: "#/components/schemas/MembershipStatus"
        inviteCode:
          type: string
          nullable: true
          description: Present when member or admin
        superAdminId:
          type: string
        adminIds:
          items:
            type: string
          type: array
        memberIds:
          items:
            type: string
          type: array
        pendingMemberIds:
          items:
            type: string
          type: array
          description: Present when admin only
        createdBy:
          type: string
        updatedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: Set when group is soft-deleted
        deletedBy:
          type: string
          nullable: true
      required:
        - id
        - name
        - desc
        - isPublic
        - memberCount
        - membershipStatus
      type: object
      additionalProperties: false
    Group:
      description: Group model - represents a social group (full, for members/admins)
      properties:
        id:
          type: string
          description: Unique group identifier
        name:
          type: string
          description: Group name
        desc:
          type: string
          description: Group description
        thumbnail:
          type: string
          nullable: true
          description: Group thumbnail/avatar URL
        avatarSeed:
          type: string
          nullable: true
          description: DiceBear icons seed for generated avatar
        inviteCode:
          type: string
          nullable: true
          description: Unique invite code for joining the group
        isPublic:
          type: boolean
          description: Whether the group is publicly visible
        superAdminId:
          type: string
          description: ID of the group's super admin
        adminIds:
          items:
            type: string
          type: array
          description: Array of admin user IDs
        memberIds:
          items:
            type: string
          type: array
          description: Array of member user IDs
        pendingMemberIds:
          items:
            type: string
          type: array
          description: Array of pending member request user IDs
        createdBy:
          type: string
          description: ID of the user who created this group
        updatedBy:
          type: string
          description: ID of the user who last updated this group
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the group was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the group was last updated
      required:
        - id
        - name
        - desc
        - isPublic
        - superAdminId
        - adminIds
        - memberIds
        - createdBy
        - updatedBy
        - createdAt
        - updatedAt
      type: object
      additionalProperties: false
    GroupInput:
      description: Input for creating a new group
      properties:
        id:
          type: string
        name:
          type: string
        desc:
          type: string
        thumbnail:
          type: string
          nullable: true
        avatarSeed:
          type: string
          nullable: true
        inviteCode:
          type: string
          nullable: true
        isPublic:
          type: boolean
        superAdminId:
          type: string
        adminIds:
          items:
            type: string
          type: array
        memberIds:
          items:
            type: string
          type: array
        createdBy:
          type: string
      required:
        - id
        - name
        - desc
        - isPublic
        - superAdminId
        - createdBy
      type: object
      additionalProperties: false
    GroupUpdate:
      description: Input for updating a group
      properties:
        name:
          type: string
        desc:
          type: string
        thumbnail:
          type: string
          nullable: true
        avatarSeed:
          type: string
          nullable: true
        isPublic:
          type: boolean
        superAdminId:
          type: string
        adminIds:
          items:
            type: string
          type: array
        memberIds:
          items:
            type: string
          type: array
        updatedBy:
          type: string
      required:
        - updatedBy
      type: object
      additionalProperties: false
    MembershipRequestAction:
      description: Membership request action
      properties:
        userId:
          type: string
          description: User ID of the member request
        action:
          type: string
          enum:
            - approve
            - reject
          description: "Action to take: approve or reject"
      required:
        - userId
        - action
      type: object
      additionalProperties: false
    Record_string.string_:
      properties: {}
      additionalProperties:
        type: string
      type: object
      description: Construct a type with a set of properties K of type T
...
  securitySchemes: {}
info:
  title: BoltUp API
  version: 1.0.0
  description: REST API for BoltUp - A social event planning platform for groups
  license:
    name: ISC
  contact:
    name: BoltUp Team
paths:
...
  /groups:
    get:
      operationId: GetGroups
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/GroupScoped"
                type: array
      description: Get all groups
      summary: Retrieves a list of groups with info scoped by user's membership
        status. Requires userId.
      tags:
        - Groups
      security: []
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
        - description: When true, includes soft-deleted groups where user is superadmin.
          in: query
          name: includeDeleted
          required: false
          schema:
            type: boolean
    post:
      operationId: CreateGroup
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
      description: Create a new group
      summary: Creates a new group with initial members
      tags:
        - Groups
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
  /groups/{id}:
    get:
      operationId: GetGroup
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupScoped"
      description: Get group by ID
      summary: Retrieves a single group with info scoped by user's membership status.
        Requires userId.
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
    put:
      operationId: UpdateGroup
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupScoped"
      description: Update a group
      summary: Updates an existing group. Requires admin.
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
    delete:
      operationId: DeleteGroup
      responses:
        "204":
          description: No Content
      description: Hard-delete a group
      summary: Permanently removes a group and all its data. Superadmin only.
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
  /groups/{id}/members:
    get:
      operationId: GetGroupMembers
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/User"
                type: array
      description: Get group members
      summary: Retrieves all members of a specific group. Requires caller to be a
        member.
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
  /groups/{id}/soft-delete:
    post:
      operationId: SoftDeleteGroup
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
      description: Soft-delete a group
      summary: Marks a group as deleted. Superadmin only.
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                userId:
                  type: string
              required:
                - userId
              type: object
  /groups/{id}/recover:
    post:
      operationId: RecoverGroup
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
      description: Recover a soft-deleted group
      summary: Restores a soft-deleted group. Superadmin only.
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                userId:
                  type: string
              required:
                - userId
              type: object
  /groups/join-by-code:
    post:
      operationId: JoinByInviteCode
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    enum:
                      - joined
                      - pending
                  groupName:
                    type: string
                  success:
                    type: boolean
                required:
                  - status
                  - groupName
                  - success
                type: object
      description: Join a group by invite code
      summary: Join or request to join a group using its invite code
      tags:
        - Groups
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                userId:
                  type: string
                inviteCode:
                  type: string
              required:
                - userId
                - inviteCode
              type: object
  /groups/{id}/leave:
    post:
      operationId: LeaveGroup
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
      description: Leave a group
      summary: Remove the current user from the group. Superadmin cannot leave.
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                userId:
                  type: string
              required:
                - userId
              type: object
  /groups/{id}/join:
    post:
      operationId: JoinGroup
      responses:
        "200":
          description: OK
      description: Join a group
      summary: Join a public group (immediate) or request to join a private group
        (pending)
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                userId:
                  type: string
              required:
                - userId
              type: object
  /groups/{id}/requests/pending:
    get:
      operationId: GetPendingRequests
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/User"
                type: array
      description: Get pending membership requests
      summary: Retrieves pending requests for a group. Requires admin.
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
  /groups/{id}/members/{memberId}/remove:
    post:
      operationId: RemoveMember
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
      description: Remove a member from the group
      summary: Admin removes a member. Cannot remove superadmin.
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: memberId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                performedBy:
                  type: string
              required:
                - performedBy
              type: object
  /groups/{id}/members/{memberId}/role:
    put:
      operationId: SetMemberRole
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
      description: Set a member's role (admin or member)
      summary: Admin sets a member's role. Cannot change superadmin.
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: memberId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                role:
                  type: string
                  enum:
                    - admin
                    - member
                performedBy:
                  type: string
              required:
                - role
                - performedBy
              type: object
  /groups/{id}/superadmin:
    put:
      operationId: SetSuperAdmin
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
      description: Transfer superadmin role to another member
      summary: Superadmin transfers ownership to an admin or member.
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                userId:
                  type: string
                performedBy:
                  type: string
              required:
                - userId
                - performedBy
              type: object
  /groups/{id}/requests/handle:
    post:
      operationId: HandleMembershipRequest
      responses:
        "204":
          description: No content
      description: Handle membership request
      summary: Approve or reject a membership request. Requires admin.
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipRequestAction"
  /groups/{id}/members/{userId}/color:
    put:
      operationId: UpdateMemberColor
      responses:
        "204":
          description: No content
      description: Update user's color preference for a group
      summary: Sets the user's custom color for a specific group
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                colorHex:
                  type: string
              required:
                - colorHex
              type: object
    get:
      operationId: GetMemberColor
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  colorHex:
                    type: string
                    nullable: true
                required:
                  - colorHex
                type: object
      description: Get user's color preference for a group
      summary: Retrieves the user's custom color for a specific group
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
  /groups/members/{userId}/colors:
    get:
      operationId: GetAllMemberColors
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Record_string.string_"
      description: Get user's color preferences for all their groups
      summary: Retrieves all group color preferences for a user
      tags:
        - Groups
      security: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
...
servers:
  - url: http://localhost:3000/api